Complete project
Completed project, finished Call to action and copyright. Added some minor tweaks to the header logo.

List of issues I have been unable to solve. I'm not sure if this is supposed to flex with everything. Currently all sections don't flex
together, but I have the page laid out in the aspect ratio of the image.

I currently learned how to implement flexbox, how to google and troubleshoot specific issues with layouts, and I got to experiment
with the box model!

I'm not satisfied with this, I might come back to this project later to redo and customize it to be something for me, but for now I feel
like I have a decent grasp on how html and css are laid out. It's only improvements from here!